First Version of Keyboard-only movement
We might have to change things to be less realistic and more video-gamey, such as speed and rotation clamping, and maybe the "back" key just turns the ship around to de-accel to a complete stop (see Endless Sky). This code is short but messy, and has no comments. That will come later with added features.